Started implementation of a pure python backend which does kind of a graph relaxation.
I have not read any literature on the approach but rather tried to
apply the principles used for neural networks to graph slam.

The basic idea is to "simply" relax the graph by locally reducing the error.
This should have several beneficial properties:
1) The graph can actually consist out of multiple not connected subgraphs
2) The iterative smoothing can have a hard realtime deadline (sacrificing precision)
3) The approach is robust to graphs that make a conventional backend (like gtsam)
numerically unstable, because it can use gradient clipping and still converge
at the cost of convergence speed if it has gradient clipping.
4) As the iterative approach is inspired by neural networks, more sophisticated
optimizers using momentum could be used as well. => escape local minima